Move the two string replacement functions to the common multiplatform module and add documentation to the functions
